refactor : Gamer, Card 패키지 분리
