+++ src/main/java/domain/BlackJackGame.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.gamer.Dealer;
+import domain.gamer.Gamer;
+import domain.gamer.Player;
 import java.util.LinkedHashMap;
 import java.util.Map;
