+++ src/main/java/domain/Decks.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.card.Card;
 import java.util.Collections;
 import java.util.List;
 import java.util.Stack;
