+++ src/main/java/domain/DecksGenerator.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.card.Card;
 import java.util.Stack;
 
 public interface DecksGenerator {
