+++ src/main/java/domain/PlayerResults.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.gamer.Player;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
