+++ src/main/java/domain/gamer/Dealer.java
@@ -1,16 +1,15 @@
-package domain;
+package domain.gamer;
 
 public class Dealer extends Gamer {
     private static final String DEALER_NAME = "딜러";
     private static final int STAY_CONDITION = 17;
 
-    // TODO : name 필드 제거 필요
     public Dealer() {
         super(new Name(DEALER_NAME));
     }
 
     @Override
-    boolean isStay() {
+    public boolean isStay() {
         return hand.sum() >= STAY_CONDITION;
     }
 }
