+++ src/main/java/domain/gamer/Gamer.java
@@ -1,17 +1,18 @@
-package domain;
+package domain.gamer;
 
+import domain.card.Card;
 import java.util.List;
 
 public abstract class Gamer {
-    Name name;
-    Hand hand;
+    protected Name name;
+    protected Hand hand;
 
     public Gamer(final Name name) {
         this.name = name;
         this.hand = new Hand();
     }
 
-    abstract boolean isStay();
+    public abstract boolean isStay();
 
     public void hit(final Card card) {
         hand.add(card);
@@ -32,7 +33,6 @@ public abstract class Gamer {
     public List<Card> getHand() {
         return hand.getCards();
     }
-
 
     public Name getName() {
         return name;
