+++ src/main/java/domain/gamer/Hand.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.gamer;
 
+import domain.card.Card;
+import domain.card.Rank;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
