+++ src/test/java/domain/GamersTest.java
@@ -3,6 +3,9 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.gamer.Dealer;
+import domain.gamer.Name;
+import domain.gamer.Player;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
